requirements.txt
opencv-python
numpy
tqdm
matplotlib
pandas
h5py
rich
colour-science
pandas
openpyxl
torch